+++ src/types/plugins/modal/popover.d.ts
@@ -1,0 +1,16 @@
+export default Popover;
+declare class Popover extends CustomElement {
+    static get properties(): {
+        title: {
+            type: StringConstructor;
+        };
+        text: {
+            type: StringConstructor;
+        };
+    };
+    title: any;
+    text: any;
+    render(): import("lit").TemplateResult<1>;
+}
+import { CustomElement } from "shared/components/element.js";
+//# sourceMappingURL=popover.d.ts.map