+++ src/types/shared/constants.d.ts
@@ -1,2 +1,20 @@
 export const VIEW_PLUGINS: string[];
+export namespace PRETTY_CHAT_STATUS {
+    let offline: any;
+    let unavailable: any;
+    let xa: any;
+    let away: any;
+    let dnd: any;
+    let chat: any;
+    let online: any;
+}
+export type PrettyChatStatus = {
+    offline: string;
+    unavailable: string;
+    xa: string;
+    away: string;
+    dnd: string;
+    chatty: string;
+    online: string;
+};
 //# sourceMappingURL=constants.d.ts.map